Move all Playwright config & setup into /config/playwright
