+++ config/playwright/playwright-db.ts
@@ -1,8 +1,8 @@
 import mongoose from 'mongoose';
 
-import { user1, user1PersonalWorkspace } from '../config/mockTestData';
-import { UserModel } from '../src/database/models/user-model';
-import { WorkspaceModel } from '../src/database/models/workspace-model';
+import { UserModel } from '../../src/database/models/user-model';
+import { WorkspaceModel } from '../../src/database/models/workspace-model';
+import { user1, user1PersonalWorkspace } from '../mockTestData';
 import { PLAYWRIGHT_MONGODB_URI } from './playwright-test-env';
 
 export async function disconnectPlaywrightDb(): Promise<void> {
